use a transform stream and do aggregation when necessary
