+++ lib/aggregator.js
@@ -31,17 +31,5 @@ Aggregator.prototype = {
 };
 
 // Get a parent and aggregate it, then stream it
-module.exports.aggregate = function(data, ID, minZ, outputFunction) {
-  var parent = ID.substring(0, ID.length-2);
-  // Does the parent object already exist? if so, aggregate its values; if not, initialize one
-  if (parentHolder[parent]) {
-    parentHolder[parent].aggregate(data);
-  } else {
-    parentHolder[parent] = new Aggregator();
-    parentHolder[parent].initialize(parent, data, function(err, child, pID) {
-      dataEater(pID, child, minZ, outputFunction);
-      parentHolder[pID] = {}
-    });
-  }
-}
+module.exports.aggregate = {}
