+++ test/eater.test.js
@@ -2,7 +2,7 @@ var split = require('split')();
 var trizilla = require('../index');
 
 process.stdin.pipe(split)
-    .on('data', trizilla.inflate)
+    .on('data', trizilla.inflate(5))
     .pipe(process.stdout);
     //.pipe(trizilla.tiler)
     // outputs tile-sized geojsons
